requirements.txt
# Core app dependencies
streamlit>=1.36
pandas>=2.2
numpy>=1.26
matplotlib>=3.9

# PCAP parsing and network tools
pyshark>=0.6
scapy>=2.5.0

# OpenAI / LM Studio client
openai>=1.30.0

# HTTP / API requests
requests>=2.32
httpx>=0.27

# Data handling
tqdm>=4.66
python-dateutil>=2.9

# Optional: for safer YAML/JSON parsing
pyyaml>=6.0
ujson>=5.9

# Security / hashing utils
cryptography>=42.0

# Dev / Test
pytest>=8.0
pytest-cov>=4.1
plotly>=5.18.0
maxminddb-geolite2>=4.1
ruff>=0.3.0
python-whois>=0.9

# Phase 3: YARA Scanning (optional)
yara-python>=4.3.0

# Phase 3: PDF Generation (optional)
weasyprint>=60.0
Jinja2>=3.1.0
markdown>=3.5.0